Fix mistake in configuring repository in Gradle
This commit fixes a typo in the configuration of the library repository in the README file that made the code uncompilable.

The fix is simple, and just ensure that the return type from the call is always `String` and not `String?`.

Moreover, this commit fixes a tiny typo that made the usename last letter lowercase.