Use default PageProperties in convert functions
This commit simplifies the usage of the code for users by adding the default `PageProperties` value as default value for each of them. This allows the consumers of the library to use it without the need to pass the `PageProperties.Builder().build()` explicitly unless they do need to configure the conversion.